improved layout, switched to intFloat for cents

+++ public/assets/js/budget.js
@@ -17,7 +17,7 @@ fetch("/api/transaction")
 function populateTotal() {
   // reduce transaction amounts to a single total value
   let total = transactions.reduce((total, t) => {
-    return total + parseInt(t.value);
+    return total + parseFloat(t.value);
   }, 0);
 
   let totalEl = document.querySelector("#total");
